new_project_starter_kit/templates/requirements.txt
# Python Dependencies Template
# For use with Python 3.11+

# =============================================================================
# CORE FRAMEWORK
# =============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# =============================================================================
# DATABASE
# =============================================================================
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9  # PostgreSQL driver

# =============================================================================
# ENVIRONMENT & CONFIGURATION
# =============================================================================
python-dotenv==1.0.0

# =============================================================================
# HTTP CLIENTS
# =============================================================================
httpx==0.25.2  # Modern async HTTP client
requests==2.31.0  # Traditional HTTP client

# =============================================================================
# DATA PROCESSING
# =============================================================================
pandas==2.1.3  # Data manipulation
numpy==1.26.2  # Numerical computing

# =============================================================================
# UTILITIES
# =============================================================================
python-multipart==0.0.6  # Form data parsing
python-jose[cryptography]==3.3.0  # JWT tokens (if using auth)
passlib[bcrypt]==1.7.4  # Password hashing (if using auth)

# =============================================================================
# CACHING (optional)
# =============================================================================
# redis==5.0.1  # Redis client
# hiredis==2.2.3  # Fast Redis parser

# =============================================================================
# MONITORING & LOGGING (optional)
# =============================================================================
# sentry-sdk[fastapi]==1.38.0  # Error tracking

# =============================================================================
# TESTING
# =============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # Also used for testing

# =============================================================================
# CODE QUALITY
# =============================================================================
black==23.11.0  # Code formatter
flake8==6.1.0  # Linter
mypy==1.7.1  # Type checker
isort==5.12.0  # Import sorter

# =============================================================================
# DEVELOPMENT TOOLS
# =============================================================================
ipython==8.18.1  # Better REPL
watchfiles==0.21.0  # Auto-reload for development

# =============================================================================
# NOTES
# =============================================================================
# 1. Pin versions to avoid breaking changes
# 2. Use pip install -r requirements.txt to install
# 3. Keep dependencies minimal - only add what you need
# 4. Update regularly with: pip list --outdated
# 5. Generate from current env with: pip freeze > requirements.txt
# 6. For development, consider splitting into requirements-dev.txt
